refactor(system): own the advertised capacity arithmetic in the system crate
Part of #1656. `mesh/capacity.rs` derives the budget a node announces and the
itemized view behind it from a `HardwareSurvey`, but it lives in
`mesh-llm-host-runtime`, which the command crates do not depend on. Anything
outside the runtime that wants to explain the announced number has to
recompute it.

Move the calculation to `mesh_llm_system::capacity`, next to the survey it
reads and the `vram` helpers it already calls. `mesh/capacity.rs` re-exports it
under the names the mesh code uses and keeps the tests that go through a node's
startup snapshot; the tests that exercise the arithmetic directly move with it.
No line of the calculation changes.

The MiB rounding of the configured safety margin moves with it as
`safety_margin_bytes`, so the three places that rounded it separately now
agree by construction: the resolver's `effective_safety_margin_bytes`
delegates, and the built-in 2 GB default has one definition.
